requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
boto3>=1.34.0
paramiko>=3.4.0
PySocks>=1.7.1
click>=8.1.0
watchdog>=4.0.0
pillow-heif>=0.15.0
psutil>=5.9.0
scikit-image>=0.22.0
numpy>=1.26.0
